Abstracts snippet building
